7.tf_two_tower/vertex-train/trainer/requirements.txt

tensorflow-recommenders==0.5.2
tensorflow-io-gcs-filesystem==0.18.0
absl-py==0.12.0
google-api-python-client==1.12.8
tfx-bsl==1.1.1
google-cloud-bigquery==2.20.0
dill==0.3.1.1
pyarrow==2.0.0
tfx-bsl==1.1.1
tensorflow-transform==1.1.0
